requirements.txt
asgiref
defusedxml
diff-match-patch
Django==3.2
django-crispy-forms
django-import-export
django-model-utils
django-notifications-hq
django-widget-tweaks
djangorestframework
et-xmlfile
gunicorn 
jsonfield
MarkupPy
mysqlclient
numpy
odfpy
openpyxl
pandas
python-dateutil
pytz
PyYAML
setuptools
six
sqlparse
swapper
tablib
wheel
xlrd
XlsxWriter
xlwt
scipy
matplotlib
sklearn
sentry_sdk
psycopg2
xhtml2pdf
django-easy-pdf
Pillow
stripe
daphne
channels
channels_redis